v.0.2

+++ database/migrations/2017_12_16_204210_add_portion_size_to_quantity_types.php
@@ -14,7 +14,7 @@ class AddPortionSizeToQuantityTypes extends Migration
     public function up()
     {
         Schema::table('quantity_types', function (Blueprint $table) {
-            $table->float('portion_size')->nullable();
+            $table->double('portion_size')->nullable();
         });
     }
